requirements.txt
PySignal>=1.1.1
PySerial>=3.5
